LayoutTests/imported/w3c/web-platform-tests/intersection-observer/v2/inline-occlusion-expected.txt

This is the target.
This is the occluder.

PASS IntersectionObserverV2 in a single document using the implicit root, with an occluding element.
FAIL First rAF. assert_true: expected true got undefined
FAIL occluder.style.marginTop = '-10px' assert_equals: expected 2 but got 1
FAIL occluder.style.opacity = 0 assert_equals: expected 3 but got 1
